+++ examples/3d_simple_character.rs
@@ -1,3 +1,14 @@
-use examples_common::*;
+use avian3d::{PhysicsPlugins, prelude::PhysicsDebugPlugin};
+use bevy::prelude::*;
+use examples_common::ExampleCommonPlugin;
 
-fn main() {}
+fn main() -> AppExit {
+    App::new()
+        .add_plugins((
+            DefaultPlugins,
+            ExampleCommonPlugin,
+            PhysicsPlugins::default(),
+            PhysicsDebugPlugin::default(),
+        ))
+        .run()
+}
